requirements.txt
amqp==2.2.2
click==6.7
coverage==4.5.1
elasticsearch==6.2.0
Flask==1.0.2
Flask-SocketIO==3.0.0
gevent==1.3.0
greenlet==0.4.13
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
pep8==1.7.1
pika==0.11.2
pycodestyle==2.4.0
python-engineio==2.1.1
python-socketio==1.9.0
redis==2.10.6
ruamel.yaml==0.15.37
selenium==3.12.0
six==1.11.0
SQLAlchemy==1.2.7
typing==3.6.4
urllib3==1.22
vine==1.1.4
Werkzeug==0.14.1
